requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Cors==4.0.0
gunicorn==21.2.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
